requirements.txt
openai>=1.0.0
genanki>=0.14.0
pysrt>=1.1.2
python-dotenv>=1.0.0